allow the use of web workers to unpack data
