+++ packages/playground/src/utils/workers.ts
@@ -1,0 +1,9 @@
+import workerpool from 'workerpool'
+import WorkerURL from './workers/unpacker?worker&url'
+
+export const workerPool = workerpool.pool(WorkerURL, {
+    maxWorkers: 2,
+    workerOpts: {
+        type: import.meta.env.PROD ? undefined : 'module',
+    },
+})
