+++ packages/playground/src/utils/workers/unpacker.ts
@@ -1,0 +1,4 @@
+import workerpool from 'workerpool'
+import { createUnpacker } from '@towns-protocol/sdk'
+
+workerpool.worker(createUnpacker())
