+++ packages/sdk/src/index.ts
@@ -124,3 +124,4 @@ export * from './views/transforms/logObservable'
 export * from './views/transforms/membershipsTransform'
 export * from './views/transforms/spaceIdsTransform'
 export * from './views/transforms/spaceUnreadsTransform'
+export * from './unpacker'
